requirements.txt
fastapi==0.104.1
Pillow==10.1.0
Requests==2.31.0
uvicorn==0.23.2
python-multipart==0.0.6
